Evolution: Phase 33.1 - Cultural Friction & Ideological Secession Engine
- Executed Vision Audit and identified lack of macro-consequences for the Memetic/Belief Engine.
- Implemented `CulturalFrictionSystem` in `internal/systems/cultural_friction.go` integrating Phase 07 (Linguistic/Belief mechanics) natively with Phase 28 (Administrative Vassal mechanics).
- Adhered strictly to DOD constraints by caching `adminCultureData` via flat slices to avoid nested `arche-go` queries.
- Proved the "Butterfly Effect" via E2E integration test demonstrating that culturally diverging vassal villages suffer Loyalty drain, which natively cascades into `VassalRebellionSystem` triggering secession (`CountryID = 0`).
- Updated `EVOLUTION_LOG.md` and `cmd/game/main.go` accordingly.

+++ internal/render/raylib_app.go
@@ -803,7 +803,7 @@ func RunRaylibApp(buildSim BuildSimFunc) {
 
 			if len(transforms) > 0 {
 				cubeMaterial.Maps.Color = rl.Blue
-				rl.DrawMeshInstanced(cubeMesh, cubeMaterial, transforms, int32(len(transforms)))
+				rl.DrawMeshInstanced(cubeMesh, cubeMaterial, transforms, len(transforms))
 			}
 
 			// Draw possessed entity (drawn normally to easily distinct it)
